requirements.txt
streamlit
python-dotenv
azure-cognitiveservices-speech
azure-ai-textanalytics
